requirements.txt
Django>=4.2.0
google-cloud-bigquery>=3.11.0
google-generativeai>=0.2.0
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
db-dtypes>=1.0.0
whitenoise
